+++ packages/mutation-testing-metrics-scala/circe/src/test/scala/mutationtesting/DecoderTest.scala
@@ -1,0 +1,5 @@
+package mutationtesting
+
+import verify._
+
+object DecoderTest extends BasicTestSuite {}