+++ CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.3](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.2.2...v1.2.3) (2019-12-11)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.2.2](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.2.1...v1.2.2) (2019-12-04)
+
+
+### Bug Fixes
+
+* **metrics:** correct lodash import ([3c5d6d2](https://github.com/stryker-mutator/mutation-testing-elements/commit/3c5d6d28e33453801e1b137330f38b616e1910c7))
+
+
+
+
+
+## [1.2.1](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.2.0...v1.2.1) (2019-11-24)
+
+
+### Features
+
+* **elements:** allow override of top offset ([#214](https://github.com/stryker-mutator/mutation-testing-elements/issues/214)) ([b6c5c36](https://github.com/stryker-mutator/mutation-testing-elements/commit/b6c5c3605c2d761c8f31d4eebd069cfc3b4e5ed9))
+* **metrics:** make `normalizeFileNames` public ([3017700](https://github.com/stryker-mutator/mutation-testing-elements/commit/3017700a5fb47a73c91e3fd9c9b849fbb49d3520))
+
+
+
+
+
+
+# [1.2.0](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.1.1...v1.2.0) (2019-10-02)
+
+
+### Features
+
+* add mutation-testing-metrics for Scala ([#65](https://github.com/stryker-mutator/mutation-testing-elements/issues/65)) ([6e732ad](https://github.com/stryker-mutator/mutation-testing-elements/commit/6e732ad))
+* **build:** use postcss to optimize css assets ([8f24f06](https://github.com/stryker-mutator/mutation-testing-elements/commit/8f24f06))
+
+
+
+
+
 ## [1.1.1](https://github.com/stryker-mutator/mutation-testing-elements/compare/v1.1.0...v1.1.1) (2019-07-15)